# refactor extreme_learning_machine
move InputToNode to _base.py
add test_base.py
remove InputToNode tests from test_elm.py

all tests successful!
